summarize query insights in saved insights view
